requirements.txt
asgiref==3.11.1
attrs==25.4.0
black==26.1.0
click==8.3.1
Django==5.2.11
django-cors-headers==4.9.0
django-environ==0.12.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
factory_boy==3.3.3
Faker==40.4.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.12.0
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.15.1
sqlparse==0.5.5
tomli==2.4.0
typing_extensions==4.15.0
uritemplate==4.2.0
